requirements.txt
streamlit>=1.28.0
torch>=2.6.0
torchvision
gdown
Pillow
safetensors